+++ packages/components/src/components/primitives/ProgressBar/ProgressBar.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+import { Progress as ChakraProgressBar } from '@chakra-ui/react';
+import { ProgressBarProps } from './types';
+
+export const ProgressBar: FC<ProgressBarProps> = (props) => (
+  <ChakraProgressBar data-testid="cmpsr.progress-bar" {...props} />
+);